+++ CHANGELOG.md
@@ -1,19 +1,68 @@
 # CHANGELOG
 
+#### v0.3.10 - DistributedArrays Update
+DistributedArrays are implemented fully using "@sync @distributed" concept of parallelization.
+
+All the changes listed:
+
+Auxiliary.jl:
+- distChunkerPadd() created to handle the lattice with paddings -- now that we are using DistArrays the field arrays need to come with paddings on the full lattice because chunks are not defined individually; there is only one field array. (Apart from f_gl arrays which are for now kept for recording purposes)
+- updatePaddings!() is defined. It calls the individual update_Padds!() functions.
+- update_Padds!() is defined. There are two definitions one where we send in a 3-indexed DArray (for ϕ and ψ) and the second one is 5-indexed DArray for Z. This function is called inside a @sync @distributed concept. This way we can send in the full DArray "f" rather than "localpart(f)" since the caller is the master proc rather than workers(). This function then takes the DArray as an argument and using the "localpart(f)" and "@fetchfrom" updates the padding of the chunks of the DArray.
+- In addition, this whole updating paddings operation is done in another way as a version 2. This version tries to use the DArray directly (at least for the half of the operations) using the global indices for each chunk. What it tries to do is that updating localparts of field arrays by calling the DArray directly with the necessary global indices for that chunk. 
+It is abandoned as it looks like it is quite slower than the version 1 described above. 
+- All the routines for parallel-Julia version are commented out.
+
+Energy.jl:
+- Definitions of energy() ang energy_calculation() has changed to send in the DistArrays.
+- update_Paddings() function is called for updating the paddings and total energy is calculated using @sync @distributed with reduction operation (+) for summation of totalE_loc's.
+- ZED_loc definition inside energy_calculation() is removed since we define ZED now as a new field which we define them as a DistArray.
+
+Evolution.jl:
+- time_evolve() and all other functions' that use fields definitions are updated to receive DArrays.
+- Snapshot interval and openning files parts are updated slightly; some comments and some adjustments for println statements.
+- Recording of the fields part is updated slighty updated for DArray construction. We have a new function to determine the global indices of local chunks. Another change is the fecthing from localpart()'s.
+- Vortex checker is added in snapping part.
+- All evolution functions are written with @sync @distributed construction and localpart() function for each field. 
+
+FindVortex.jl:
+- Vortex finder has a commented out piece to use directly the DArray rather than ϕ_gl which is defined for recording purposes which might be changed later if we can find a way to just use the padded DArrays rather than f_gl's.
+
+IC.jl:
+- Cleaned out the unnecessary junk.
+- renormalization() and zeroPointEnergy() are updated to work with DArrays. Instead of @fetchfrom'ing we use DArray and just call what we want like a serial code.
+
+main.jl:
+- f_gl's are defined without offset. They are still defined for recording purposes for the time being. We could just record the DArrays but the fields have paddings and recording paddings seems useless. We can test the speed of this vs using f_gl's later. 
+- ZED is defined as a DArray as well but doesn't have a global version since it doesn't have a padding.
+- Recording the information about lattice/parameters part is updated; writedlm() removed, using just println().
+- initial_ZED is recorded.
+
+Next:
+We need to test speed recording f_gl() version vs DArray recording (either by removing the paddings before recording or removing paddings in the plotting)
+Also is there maybe a way to do all of these with DArray without using the @sync @distributed construct? Like the one they do in DArray documentation but optimized for memory.
+
+
+
 #### v0.3.9 - DistributedArrays
 Complete new version of parallelization using DistributedArrays.jl package
 ---Still work in progress---
 
 All the changes listed:
+
 Auxiliary.jl:
 - distChunker() created. Finds the (end) coordinates of arrays of local chunks in the global lattice based on their proc id. Same as chunker() but without the "Offset".
+
 FindVortex.jl:
 - Updated for recording the positions of vortices and anti-vortices and returns them.
+
 IC.jl:
 - localpart(f) is used to let the initial conditions for ϕ and ψ. ϕ_gl and ψ_gl is defined for matrix calculations of initial Z's. Later Z is initialized chunk-wise with the "global" omega matrices.
 - Overall indices is changed since DistibutedArrays doesn't support Offset'ting.
+
 Main.jl:
 - All the arrays are now defined with indices starting 1. Offset's are now turned off.
+
 Parameters.jl:
 - New padded lattice sizes defined for DistributedArray's: Nx_padd,Ny_padd.
 
@@ -23,19 +72,24 @@ MPI Code updated --> Z_gl and dZdt_gl are no longer defined. Should improve memo
 
 #### v0.3.6
 All the changes listed
+
 Auxiliary.jl:
 - For the parallel functions' arguments, the types are added for performance improvement. 
 - update_Padddings() returns nothing now. Before it was accidently returning the last line which had some performance hit.
 - getData functions' if statements covers everything now; else added with error statement.
+
 Evolution.jl:
+
 IC.jl:
 - Z_gl and dZdt_gl is removed from arguments and local Z's are now calcualted directly from the omega matrices.
 - Fluctuations are added for the field ϕ
 - Small synhtatic updates.
 - S_Ωzero, inv_S_Ωzero are now created first with zeros earlier, then updated with sqrt()'s etc.
+
 main.jl:
 - Z_gl and dZdt_gl is no longer created. We set local Z's directly on the workers now using 4-index omega matrices.
 - Vortex finder is called after time_evolution()
+
 Parameters.jl:
 - lx, rx are defined with Int() rather than Int.16() for performance reasons.
 - dt_half is defined here for performance.